Move Feature Flag Tests out of Remote Config Tests
And add v4 specific tests.
